+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # clancy-proxy
 
+## 0.0.5-beta.3
+
+### Patch Changes
+
+- make utls esm
+- Updated dependencies
+  - @clancyapp/shared@0.0.5-beta.3
+
+## 0.0.5-beta.2
+
+### Patch Changes
+
+- Improve LLM response view
+- Updated dependencies
+  - @clancyapp/shared@0.0.5-beta.2
+
 ## 0.0.5-beta.1
 
 ### Patch Changes