+++ packages/utls/platforms/darwin-arm64/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @clancyapp/utls-darwin-arm64
 
+## 0.0.5-beta.3
+
+### Patch Changes
+
+- make utls esm
+
+## 0.0.5-beta.2
+
+### Patch Changes
+
+- Improve LLM response view
+
 ## 0.0.5-beta.1
 
 ### Patch Changes